requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.3.4
appnope==0.1.3
asgiref==3.4.1
astroid==2.13.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.1.0
chromedriver-binary==114.0.5735.90.0
click==8.0.3
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
dnspython==2.1.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.70.0
fastapi-csrf-protect==0.2.1
fonttools==4.40.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gunicorn==20.1.0
h11==0.12.0
idna==3.3
importlib-metadata==6.8.0
ipykernel==6.22.0
ipython==8.11.0
isort==5.11.4
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
motor==2.4.0
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.2
openai==0.27.8
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.6.2
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.20
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.14.0
PyJWT==2.3.0
pylint==2.15.10
pymongo==3.12.1
Pympler==1.0.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML @ file:///private/tmp/pyyaml-20230718-6214-18vax93/PyYAML-6.0.1
pyzmq==25.0.2
referencing==0.29.1
requests==2.28.2
rich==13.4.2
rpds-py==0.8.8
selenium==4.8.2
semver==3.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
stack-data==0.6.2
starlette==0.16.0
streamlit==1.20.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.15
uvicorn==0.14.0
validators==0.20.0
wcwidth==0.2.6
webdriver-manager==3.8.5
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
